+++ src/core/gfx/load/pe-reader.js
@@ -63,13 +63,13 @@ export class PEReader {
     this.position += offset;
   }
 
-  readShort() {
+  readUint16() {
     let result = this.dataView.getUint16(this.position, true);
     this.position += 2;
     return result;
   }
 
-  readInt() {
+  readUint32() {
     let result = this.dataView.getUint32(this.position, true);
     this.position += 4;
     return result;
@@ -86,28 +86,28 @@ export class PEReader {
 
   readDirectoryEntryCount() {
     this.skip(0x0c);
-    let namedEntries = this.readShort();
-    let idEntries = this.readShort();
+    let namedEntries = this.readUint16();
+    let idEntries = this.readUint16();
     return namedEntries + idEntries;
   }
 
   readResourceDirectoryEntry() {
-    let resourceType = this.readInt();
-    let subdirectoryOffset = this.readInt();
+    let resourceType = this.readUint32();
+    let subdirectoryOffset = this.readUint32();
     return new ResourceDirectoryEntry(resourceType, subdirectoryOffset);
   }
 
   readResourceDataEntry() {
-    let offset = this.readInt();
-    let size = this.readInt();
-    let codePage = this.readInt();
-    let unused = this.readInt();
+    let offset = this.readUint32();
+    let size = this.readUint32();
+    let codePage = this.readUint32();
+    let unused = this.readUint32();
     return new ResourceDataEntry(offset, size, codePage, unused);
   }
 
   readHeader() {
     this.seek(0x3c);
-    let peHeaderAddress = this.readShort();
+    let peHeaderAddress = this.readUint16();
 
     this.skip(peHeaderAddress - 0x3c - 0x02);
     let type = this.readString(4);
@@ -116,18 +116,18 @@ export class PEReader {
     }
 
     this.skip(0x02);
-    let sections = this.readShort();
+    let sections = this.readUint16();
 
     this.skip(0x78 - 0x04 + 0x0c);
-    this.virtualAddress = this.readInt();
+    this.virtualAddress = this.readUint32();
 
     this.skip(0x6c + 0x08 + 0x04);
 
     for (let i = 0; i < sections; ++i) {
-      let checkVirtualAddress = this.readInt();
+      let checkVirtualAddress = this.readUint32();
       if (checkVirtualAddress == this.virtualAddress) {
         this.skip(0x04);
-        this.rootAddress = this.readInt();
+        this.rootAddress = this.readUint32();
         break;
       }
       this.skip(0x24);
@@ -167,16 +167,28 @@ export class PEReader {
 
     for (let entry of bitmapEntries) {
       this.seek(this.rootAddress + entry.subdirectoryOffset + 20);
-      let entrySubdirectoryOffset = this.readInt();
+      let entrySubdirectoryOffset = this.readUint32();
 
       this.seek(this.rootAddress + entrySubdirectoryOffset);
       let dataEntry = this.readResourceDataEntry();
       let start = dataEntry.offset - this.virtualAddress + this.rootAddress;
       let size = dataEntry.size;
 
-      this.seek(start + 4);
-      let width = this.readInt();
-      let height = this.readInt();
+      this.seek(start);
+
+      let headerSize = this.readUint32();
+      let width;
+      let height;
+
+      if (headerSize === 12) {
+        // BITMAPCOREHEADER: width and height are unsigned 16-bit values
+        // See: https://learn.microsoft.com/windows/win32/api/wingdi/ns-wingdi-bitmapcoreheader
+        width = this.readUint16();
+        height = this.readUint16();
+      } else {
+        width = this.readUint32();
+        height = this.readUint32();
+      }
 
       this.resourceInfo.set(
         entry.resourceType,